crypto: keystone: Added support for HW RNG
This patch adds the TRNG module in CP_ACE as a HW RNG source in Kernel. Random data generated by the hardware will be accesible to user space via the character device "/dev/hwrng". To effectively use this feature it is recommended to run "rngd" daemon (available in rng-tools package) in user space, which feeds the random number entropy pool of the kernel with random data from the HW RNG. This feature has been tested using the "rngtest" (rng-tools) program.

Signed-off-by: Sandeep Nair <sandeep_n@ti.com>